@@ -4,28 +4,8 @@ import (
 	"net/http"
 )
 
-// AppController handles the HTTP requests
-type AppController interface {
-	Healthz(w http.ResponseWriter, r *http.Request)
-	Statusz(w http.ResponseWriter, r *http.Request)
-}
-
-// appController struct holds the usecase
-type appController struct {
-	commit  string
-	version string
-}
-
-// NewAppController returns an EventController
-func NewAppController(commit string, version string) AppController {
-	return &appController{
-		commit:  commit,
-		version: version,
-	}
-}
-
 // Healthz controller
-func (ac *appController) Healthz(w http.ResponseWriter, r *http.Request) {
+func (c controller) Healthz(w http.ResponseWriter, r *http.Request) {
 	JSONResponse(
 		w,
 		http.StatusOK,
@@ -37,13 +17,13 @@ func (ac *appController) Healthz(w http.ResponseWriter, r *http.Request) {
 }
 
 // Statusz controller
-func (ac *appController) Statusz(w http.ResponseWriter, r *http.Request) {
+func (c controller) Statusz(w http.ResponseWriter, r *http.Request) {
 	JSONResponse(
 		w,
 		http.StatusOK,
 		map[string]interface{}{
-			"commit":  ac.commit,
-			"version": ac.version,
+			"commit":  c.commit,
+			"version": c.version,
 		},
 	)
 }
