@@ -1,12 +1,13 @@
-// Code generated by mockery v0.0.0-dev. DO NOT EDIT.
+// Code generated by mockery v2.14.0. DO NOT EDIT.
 
 package mocks
 
 import (
 	context "context"
 
-	model "github.com/javiertlopez/awesome/model"
 	mock "github.com/stretchr/testify/mock"
+
+	model "github.com/javiertlopez/awesome/model"
 )
 
 // Delivery is an autogenerated mock type for the Delivery type
@@ -33,4 +34,19 @@ func (_m *Delivery) GetByID(ctx context.Context, id string) (model.Video, error)
 	}
 
 	return r0, r1
+}
+
+type mockConstructorTestingTNewDelivery interface {
+	mock.TestingT
+	Cleanup(func())
+}
+
+// NewDelivery creates a new instance of Delivery. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+func NewDelivery(t mockConstructorTestingTNewDelivery) *Delivery {
+	mock := &Delivery{}
+	mock.Mock.Test(t)
+
+	t.Cleanup(func() { mock.AssertExpectations(t) })
+
+	return mock
 }
