@@ -1,12 +1,13 @@
-// Code generated by mockery v0.0.0-dev. DO NOT EDIT.
+// Code generated by mockery v2.14.0. DO NOT EDIT.
 
 package mocks
 
 import (
 	context "context"
 
-	model "github.com/javiertlopez/awesome/model"
 	mock "github.com/stretchr/testify/mock"
+
+	model "github.com/javiertlopez/awesome/model"
 )
 
 // Ingestion is an autogenerated mock type for the Ingestion type
@@ -33,4 +34,19 @@ func (_m *Ingestion) Create(ctx context.Context, anyVideo model.Video) (model.Vi
 	}
 
 	return r0, r1
+}
+
+type mockConstructorTestingTNewIngestion interface {
+	mock.TestingT
+	Cleanup(func())
+}
+
+// NewIngestion creates a new instance of Ingestion. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+func NewIngestion(t mockConstructorTestingTNewIngestion) *Ingestion {
+	mock := &Ingestion{}
+	mock.Mock.Test(t)
+
+	t.Cleanup(func() { mock.AssertExpectations(t) })
+
+	return mock
 }
