@@ -4,32 +4,30 @@ import (
 	"context"
 
 	"github.com/javiertlopez/awesome/model"
-	"github.com/javiertlopez/awesome/repository"
+	"github.com/sirupsen/logrus"
 )
 
-// Delivery usecase
-type Delivery interface {
-	GetByID(ctx context.Context, id string) (model.Video, error)
-}
-
 type delivery struct {
-	assets repository.AssetRepo
-	videos repository.VideoRepo
+	assets Assets
+	videos Videos
+	logger *logrus.Logger
 }
 
-// NewDelivery returns the usecase implementation
-func NewDelivery(
-	a repository.AssetRepo,
-	v repository.VideoRepo,
-) Delivery {
-	return &delivery{
+// Delivery returns the usecase implementation
+func Delivery(
+	a Assets,
+	v Videos,
+	l *logrus.Logger,
+) delivery {
+	return delivery{
 		assets: a,
 		videos: v,
+		logger: l,
 	}
 }
 
 // GetByID methods
-func (u *delivery) GetByID(ctx context.Context, id string) (model.Video, error) {
+func (u delivery) GetByID(ctx context.Context, id string) (model.Video, error) {
 	response, err := u.videos.GetByID(ctx, id)
 
 	if err != nil {
@@ -39,12 +37,14 @@ func (u *delivery) GetByID(ctx context.Context, id string) (model.Video, error) 
 	// If video document contains an Asset ID, retrieve the information
 	if response.Asset != nil {
 		asset, err := u.assets.GetByID(ctx, response.Asset.ID)
-		if err == nil {
-			response.Asset = nil
-			response.Poster = asset.Poster
-			response.Thumbnail = asset.Thumbnail
-			response.Sources = asset.Sources
+		if err != nil {
+			u.logger.Error(err)
+			return response, nil
 		}
+
+		response.Poster = asset.Poster
+		response.Thumbnail = asset.Thumbnail
+		response.Sources = asset.Sources
 	}
 
 	return response, nil
