@@ -6,13 +6,20 @@ import (
 	"reflect"
 	"testing"
 
-	mocks "github.com/javiertlopez/awesome/mocks/repository"
-	"github.com/javiertlopez/awesome/model"
 	"github.com/stretchr/testify/mock"
+
+	"github.com/javiertlopez/awesome/model"
+	"github.com/javiertlopez/awesome/usecase/mocks"
 )
+
+// Generate mocks
+// mockery --keeptree --name=Assets --dir=usecase --output=usecase/mocks
 
 func Test_ingestion_Create(t *testing.T) {
 	uuid := "4e5bf8f2-9c50-4576-b9d4-1d1fd0705885"
+	asset := model.Asset{
+		ID: uuid,
+	}
 	expected := model.Video{
 		Title:       "Some Might Say",
 		Description: "(What's the Story) Morning Glory?",
@@ -97,21 +104,21 @@ func Test_ingestion_Create(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			ar := &mocks.AssetRepo{}
-			vr := &mocks.VideoRepo{}
-			u := &ingestion{
-				ar,
-				vr,
+			assets := &mocks.Assets{}
+			videos := &mocks.Videos{}
+			usecase := &ingestion{
+				assets,
+				videos,
 			}
 
 			if tt.wantErr {
-				ar.On("Create", tt.args.ctx, tt.args.anyVideo.SourceURL, mock.Anything).Return(uuid, errors.New("failed"))
-				vr.On("Create", tt.args.ctx, tt.args.anyVideo).Return(model.Video{}, errors.New("failed"))
+				assets.On("Create", tt.args.ctx, tt.args.anyVideo.SourceURL, mock.Anything).Return(model.Asset{}, errors.New("failed"))
+				videos.On("Create", tt.args.ctx, tt.args.anyVideo).Return(model.Video{}, errors.New("failed"))
 			} else {
-				ar.On("Create", tt.args.ctx, tt.args.anyVideo.SourceURL, mock.Anything).Return(uuid, nil)
-				vr.On("Create", tt.args.ctx, mock.Anything).Return(tt.want, nil)
+				assets.On("Create", tt.args.ctx, tt.args.anyVideo.SourceURL, mock.Anything).Return(asset, nil)
+				videos.On("Create", tt.args.ctx, mock.Anything).Return(tt.want, nil)
 			}
-			got, err := u.Create(tt.args.ctx, tt.args.anyVideo)
+			got, err := usecase.Create(tt.args.ctx, tt.args.anyVideo)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("ingestion.Create() error = %v, wantErr %v", err, tt.wantErr)
 				return
