@@ -1,4 +1,4 @@
-// Code generated by mockery v0.0.0-dev. DO NOT EDIT.
+// Code generated by mockery v2.14.0. DO NOT EDIT.
 
 package mocks
 
@@ -9,20 +9,20 @@ import (
 	mock "github.com/stretchr/testify/mock"
 )
 
-// AssetRepo is an autogenerated mock type for the AssetRepo type
-type AssetRepo struct {
+// Assets is an autogenerated mock type for the Assets type
+type Assets struct {
 	mock.Mock
 }
 
 // Create provides a mock function with given fields: ctx, source, public
-func (_m *AssetRepo) Create(ctx context.Context, source string, public bool) (string, error) {
+func (_m *Assets) Create(ctx context.Context, source string, public bool) (model.Asset, error) {
 	ret := _m.Called(ctx, source, public)
 
-	var r0 string
-	if rf, ok := ret.Get(0).(func(context.Context, string, bool) string); ok {
+	var r0 model.Asset
+	if rf, ok := ret.Get(0).(func(context.Context, string, bool) model.Asset); ok {
 		r0 = rf(ctx, source, public)
 	} else {
-		r0 = ret.Get(0).(string)
+		r0 = ret.Get(0).(model.Asset)
 	}
 
 	var r1 error
@@ -36,7 +36,7 @@ func (_m *AssetRepo) Create(ctx context.Context, source string, public bool) (st
 }
 
 // GetByID provides a mock function with given fields: ctx, id
-func (_m *AssetRepo) GetByID(ctx context.Context, id string) (model.Asset, error) {
+func (_m *Assets) GetByID(ctx context.Context, id string) (model.Asset, error) {
 	ret := _m.Called(ctx, id)
 
 	var r0 model.Asset
@@ -54,4 +54,19 @@ func (_m *AssetRepo) GetByID(ctx context.Context, id string) (model.Asset, error
 	}
 
 	return r0, r1
+}
+
+type mockConstructorTestingTNewAssets interface {
+	mock.TestingT
+	Cleanup(func())
+}
+
+// NewAssets creates a new instance of Assets. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+func NewAssets(t mockConstructorTestingTNewAssets) *Assets {
+	mock := &Assets{}
+	mock.Mock.Test(t)
+
+	t.Cleanup(func() { mock.AssertExpectations(t) })
+
+	return mock
 }
