@@ -1,4 +1,4 @@
-// Code generated by mockery v0.0.0-dev. DO NOT EDIT.
+// Code generated by mockery v2.14.0. DO NOT EDIT.
 
 package mocks
 
@@ -9,13 +9,13 @@ import (
 	mock "github.com/stretchr/testify/mock"
 )
 
-// VideoRepo is an autogenerated mock type for the VideoRepo type
-type VideoRepo struct {
+// Videos is an autogenerated mock type for the Videos type
+type Videos struct {
 	mock.Mock
 }
 
 // Create provides a mock function with given fields: ctx, anyVideo
-func (_m *VideoRepo) Create(ctx context.Context, anyVideo model.Video) (model.Video, error) {
+func (_m *Videos) Create(ctx context.Context, anyVideo model.Video) (model.Video, error) {
 	ret := _m.Called(ctx, anyVideo)
 
 	var r0 model.Video
@@ -36,7 +36,7 @@ func (_m *VideoRepo) Create(ctx context.Context, anyVideo model.Video) (model.Vi
 }
 
 // GetByID provides a mock function with given fields: ctx, id
-func (_m *VideoRepo) GetByID(ctx context.Context, id string) (model.Video, error) {
+func (_m *Videos) GetByID(ctx context.Context, id string) (model.Video, error) {
 	ret := _m.Called(ctx, id)
 
 	var r0 model.Video
@@ -54,4 +54,19 @@ func (_m *VideoRepo) GetByID(ctx context.Context, id string) (model.Video, error
 	}
 
 	return r0, r1
+}
+
+type mockConstructorTestingTNewVideos interface {
+	mock.TestingT
+	Cleanup(func())
+}
+
+// NewVideos creates a new instance of Videos. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+func NewVideos(t mockConstructorTestingTNewVideos) *Videos {
+	mock := &Videos{}
+	mock.Mock.Test(t)
+
+	t.Cleanup(func() { mock.AssertExpectations(t) })
+
+	return mock
 }
